refactor(progress): extract duplicate filtering logic to shared utility function
